Fix ApiKey detection so it works on Linux
The Steam Compatibility Layer creates a virtual C: drive for each
installed game. Which means the Tsw6Api Key is in a different location
than on Windows and the API cannot find it.

This patch changes the search logic to detect if we are running on Linux
and if so change the search location so it finds the Key.